Distribute Bonuses Problem
You are the manager of a number of employees who all sit in a row. The CEO would like to give all of your employees bonuses, but since the company did not perform so well this year the CEO would like to keep the bonuses to a minimum.

The rule of the distribution:
- Each Employee begins with a bonus factor of 1x.
- For each employee, if they perform better than the persoin sitting next to them, the employee is given a +1 higher bonus (and up to +2 if they perform better than both people to their sides).

Given a list of employee's performance, find the bonuses each employee should get.

Example:
Input: [1,2,3,2,3,5,1]
Output: [1,2,3,1,2,3,1]