H-Index solver
H-Index

The H-index is a metric that attempt to measure the productivity and citation impact of the publication of a scholar. The definition of H-index is if a scholar has a least h of their papers cited h times.

Given a list of publications of the number of citations a scholar has, find their h-index.

Example:
Input: [3,5,0,1,3]
Output: 3
Explanation:
there are 3 publications with 3 or more citations, hence the h-index is 3.